Added deferred example

+++ src/camera.rs
@@ -20,21 +20,21 @@ pub struct Camera {
     target: glm::Vec3,
     z_near: f32,
     z_far: f32,
-    width: u32,
-    height: u32,
+    width: usize,
+    height: usize,
 }
 
 
 impl Camera
 {
-    pub fn create(gl: &gl::Gl, position: glm::Vec3, target: glm::Vec3, width: u32, height: u32) -> Result<Camera, Error>
+    pub fn create(gl: &gl::Gl, position: glm::Vec3, target: glm::Vec3, width: usize, height: usize) -> Result<Camera, Error>
     {
         let mut camera = Camera { gl: gl.clone(), position, target, z_near: 0.1, z_far: 1000.0, width: width, height: height };
         camera.set_screen_size(width, height);
         Ok(camera)
     }
 
-    pub fn set_screen_size(&mut self, width: u32, height: u32)
+    pub fn set_screen_size(&mut self, width: usize, height: usize)
     {
         unsafe {
             self.gl.Viewport(0, 0, width as i32, height as i32);
